pyproject.toml
[project]
name = "worship-ppt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "pyperclip>=1.11.0",
    "python-pptx>=1.0.2",
    "streamlit>=1.52.2",
]
